mostly moving things around
