+++ .gitignore
@@ -3,4 +3,4 @@
 *.iml
 .idea
 data/
-randtests/
+cmd/randtests/
