+++ internal/mods/pingmod/ping.go
@@ -1,0 +1,168 @@
+package pingmod
+
+import (
+	"fmt"
+	"github.com/intrntsrfr/meido/internal/base"
+	"math/rand"
+	"sync"
+	"time"
+)
+
+// PingMod represents the ping mod
+type PingMod struct {
+	sync.Mutex
+	name         string
+	commands     map[string]*base.ModCommand
+	allowedTypes base.MessageType
+	allowDMs     bool
+}
+
+// New returns a new PingMod.
+func New(n string) base.Mod {
+	return &PingMod{
+		name:         n,
+		commands:     make(map[string]*base.ModCommand),
+		allowedTypes: base.MessageTypeCreate,
+		allowDMs:     true,
+	}
+}
+
+// Name returns the name of the mod.
+func (m *PingMod) Name() string {
+	return m.name
+}
+
+// Save saves the mod state to a file.
+func (m *PingMod) Save() error {
+	return nil
+}
+
+// Load loads the mod state from a file.
+func (m *PingMod) Load() error {
+	return nil
+}
+
+// Passives returns the mod passives.
+func (m *PingMod) Passives() []*base.ModPassive {
+	return []*base.ModPassive{}
+}
+
+// Commands returns the mod commands.
+func (m *PingMod) Commands() map[string]*base.ModCommand {
+	return m.commands
+}
+
+// AllowedTypes returns the allowed MessageTypes.
+func (m *PingMod) AllowedTypes() base.MessageType {
+	return m.allowedTypes
+}
+
+// AllowDMs returns whether the mod allows DMs.
+func (m *PingMod) AllowDMs() bool {
+	return m.allowDMs
+}
+
+// Hook will hook the Mod into the Bot.
+func (m *PingMod) Hook(b *base.Bot) error {
+	err := m.Load()
+	if err != nil {
+		return err
+	}
+
+	rand.Seed(time.Now().Unix())
+
+	m.RegisterCommand(NewPingCommand(m))
+	m.RegisterCommand(NewMonkeyCommand(m))
+
+	return nil
+}
+
+// RegisterCommand registers a ModCommand to the Mod
+func (m *PingMod) RegisterCommand(cmd *base.ModCommand) {
+	m.Lock()
+	defer m.Unlock()
+	if _, ok := m.commands[cmd.Name]; ok {
+		panic(fmt.Sprintf("command '%v' already exists in %v", cmd.Name, m.Name()))
+	}
+	m.commands[cmd.Name] = cmd
+}
+
+// NewPingCommand returns a new ping command.
+func NewPingCommand(m *PingMod) *base.ModCommand {
+	return &base.ModCommand{
+		Mod:           m,
+		Name:          "ping",
+		Description:   "Checks ping",
+		Triggers:      []string{"m?ping"},
+		Usage:         "m?ping",
+		Cooldown:      2,
+		RequiredPerms: 0,
+		RequiresOwner: false,
+		AllowedTypes:  base.MessageTypeCreate,
+		AllowDMs:      true,
+		Enabled:       true,
+		Run:           m.pingCommand,
+	}
+}
+
+func (m *PingMod) pingCommand(msg *base.DiscordMessage) {
+	if msg.LenArgs() < 1 {
+		return
+	}
+
+	startTime := time.Now()
+
+	first, err := msg.Reply("Ping")
+	if err != nil {
+		return
+	}
+
+	now := time.Now()
+	discordLatency := now.Sub(startTime)
+	botLatency := now.Sub(msg.TimeReceived)
+
+	msg.Sess.ChannelMessageEdit(msg.Message.ChannelID, first.ID,
+		fmt.Sprintf("Pong!\nDiscord delay: %s\nBot delay: %s", discordLatency, botLatency))
+}
+
+// NewMonkeyCommand returns a new monkey command.
+func NewMonkeyCommand(m *PingMod) *base.ModCommand {
+	return &base.ModCommand{
+		Mod:           m,
+		Name:          "monkey",
+		Description:   "Monkey",
+		Triggers:      []string{"m?monkey", "m?monke", "m?monki", "m?monky"},
+		Usage:         "m?monkey",
+		Cooldown:      0,
+		RequiredPerms: 0,
+		RequiresOwner: false,
+		AllowedTypes:  base.MessageTypeCreate,
+		AllowDMs:      true,
+		Enabled:       true,
+		Run:           m.monkeyCommand,
+	}
+}
+
+func (m *PingMod) monkeyCommand(msg *base.DiscordMessage) {
+	msg.Reply(monkeys[rand.Intn(len(monkeys))])
+}
+
+var monkeys = []string{
+	"🐒",
+	"🐒💨",
+	"🔫🐒",
+	"🎷🐒",
+	"\U0001F9FB🖊️🐒",
+	"🐒🚿",
+	"🐒\n🚽",
+	"🍌🐒",
+	"🥁🐒",
+	"\U0001FA98🐒",
+	"🏓🐒",
+	"🏸🐒",
+	"🏀🐒",
+	"🔨🐒",
+	"⛏️🐒",
+	"\U0001FAA0🐒",
+	"👑\n🐒",
+}
